Created a workaround for Volley and handling response codes + others
The server now returns the response code inside the body of the response.
Alongside the response code, the response holds also the company name and vehicle status
which get passed to the next screen for displaying purposes.
Added a new layout for the details on the next screen and added the vehicle status